refactor: add a di package to each module and remove a separate di module
