--- core/di/build.gradle.kts
@@ -1,74 +1,0 @@
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.ksp)
-    alias(libs.plugins.hilt.plugin)
-}
-
-android {
-    namespace = "ru.itis.androidhomework"
-    compileSdk = libs.versions.compileSdk.get().toInt()
-
-    defaultConfig {
-        minSdk = libs.versions.minSdk.get().toInt()
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-
-    buildFeatures {
-        buildConfig = true
-    }
-    
-    defaultConfig {
-        buildConfigField(
-            "String",
-            "OPEN_TRIP_MAP_BASE_URL",
-            "\"https://api.opentripmap.com/0.1/ru/\""
-        )
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = "11"
-    }
-}
-
-dependencies {
-
-    implementation(project(":core:data"))
-    implementation(project(":core:domain"))
-
-    implementation(libs.hilt)
-    ksp(libs.hilt.compiler)
-    implementation(libs.okhttp)
-    implementation(libs.retrofit.gson.converter)
-    implementation(libs.http.loggin.interceptor)
-
-    implementation(libs.room)
-    ksp(libs.room.ksp)
-    implementation(libs.room.ktx)
-
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
-    implementation(libs.material)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.espresso.core)
-
-    implementation(libs.navigation.fragment)
-    implementation(libs.navigation.ui)
-}
