--- core/di/src/main/java/ru/itis/androidhomework/di/module/NavigationModule.kt
@@ -1,26 +1,0 @@
-package ru.itis.androidhomework.di.module
-//
-//import dagger.Binds
-//import dagger.Module
-//import dagger.hilt.InstallIn
-//import dagger.hilt.components.SingletonComponent
-//import ru.itis.androidhomework.presentation.navigation.nav.Nav
-//import ru.itis.androidhomework.presentation.navigation.nav.NavMain
-//import ru.itis.androidhomework.presentation.navigation.navImpl.NavImpl
-//import ru.itis.androidhomework.presentation.navigation.navImpl.NavMainImpl
-//import javax.inject.Singleton
-//
-//
-//@Module
-//@InstallIn(SingletonComponent::class)
-//interface NavigationModule {
-//
-//    @Binds
-//    @Singleton
-//    fun bindNavToImpl(impl: NavImpl): Nav
-//
-//    @Binds
-//    @Singleton
-//    fun bindNavMainToImpl(impl: NavMainImpl): NavMain
-//
-//}
