+++ settings.gradle.kts
@@ -26,7 +26,6 @@ include(":data")
 include(":core:data")
 include(":core:domain")
 include(":core:base")
-include(":core:di")
 include(":core:navigation")
 include(":feature:auth")
 include(":feature:auth")
